[Fix] 튜토리얼 팝업 위치 홈 -> KUMAP
